eir: Generalize booting
frigg: Add constants for PAT detection
